Step deleting functionallity.
