+++ client/src/components/recipes/CreateRecipeForm.svelte
@@ -14,34 +14,6 @@
   let ingredients = [];
   let unitsOfMeasurement = [];
 
-  $: steps = recipeForm.steps;
-
-  const addStep = e => {
-    const newStep = {
-      description: '',
-      time: '',
-      orderNumber: recipeForm.steps.length + 1,
-      ingredients: [],
-    };
-
-    recipeForm.steps = [...recipeForm.steps, newStep];
-  };
-
-  const moveStepUp = e => {
-    moveHelper(e.detail);
-  };
-
-  const moveStepDown = e => {
-    // To move step down, it's like moving the next step up, same thing - so we use index of next step, which is e.detail + 1
-    moveHelper(e.detail + 1);
-  };
-
-  const moveHelper = stepIndex => {
-    let newSteps = recipeForm.steps.slice(0, stepIndex - 1);
-    newSteps.push(steps[stepIndex], steps[stepIndex - 1], ...steps.slice(stepIndex + 1));
-    recipeForm.steps = [...newSteps];
-  };
-
   const handleFormSubmit = e => {
     console.log('Submit');
     console.log(recipeForm);
@@ -50,22 +22,10 @@
   onMount(async () => {
     // TODO: currently, if fetch fails, categories are set to [], but in the future this should inform user about error!
     // TODO: currently, if fetch fails, ingredients and units are set to [], but in the future this should inform user about error! (same story as categories)
-    categories = await categoriesApi.fetchAll().catch(() => [
-      { id: 0, name: 'meat' },
-      { id: 1, name: 'vegan' },
-    ]);
-    ingredients = await ingredientsApi.fetchAll().catch(() => [
-      { id: 0, name: 'corn' },
-      { id: 1, name: 'cream' },
-    ]);
-    unitsOfMeasurement = await unitsOfMeasurementApi.fetchAll().catch(() => [
-      { id: 0, name: 'kilogram' },
-      { id: 1, name: 'liter' },
-    ]);
+    categories = await categoriesApi.fetchAll().catch(() => []);
+    ingredients = await ingredientsApi.fetchAll().catch(() => []);
+    unitsOfMeasurement = await unitsOfMeasurementApi.fetchAll().catch(() => []);
   });
-
-  // This will force user to have at least one step for each recipe
-  addStep();
 </script>
 
 <label class="item" for="name">Recipe name: </label>
@@ -83,10 +43,7 @@
   {/each}
 </select>
 <CreateStepsList
-  on:add-step={addStep}
-  on:move-up={moveStepUp}
-  on:move-down={moveStepDown}
-  {steps}
+  bind:steps={recipeForm.steps}
   {ingredients}
   {unitsOfMeasurement}
 />
